fetch/metadata/generated/worker-dedicated-constructor-sec-fetch-mode.https.sub.html

<!DOCTYPE html>
<!-- This test was procedurally generated. Please do not modify it directly. -->
<html lang="en">
  <meta charset="utf-8">
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
  <script src="/fetch/metadata/resources/helper.sub.js"></script>
  <script type="module">
  'use strict';
  function induceRequest(url, options) {
    return new Promise((resolve, reject) => {
      const worker = new Worker(url, options);
      worker.onmessage = resolve;
      worker.onerror = reject;
    });
  }

  promise_test(() => {
    const key = '{{uuid()}}';
    const url = makeRequestURL(
      key,
      [],
      { mime: 'application/javascript', body: 'postMessage("")' }
    );

    return induceRequest(url)
      .then(() => retrieve(key))
      .then((headers) => {
          assert_own_property(headers, 'sec-fetch-mode');
          assert_equals(headers['sec-fetch-mode'], 'same-origin');
        });
  }, 'classic script');

  promise_test(() => {
    const key = '{{uuid()}}';
    const url = makeRequestURL(
      key,
      [],
      { mime: 'application/javascript', body: 'postMessage("")' }
    );

    return induceRequest(url, {"type": "module"})
      .then(() => retrieve(key))
      .then((headers) => {
          assert_own_property(headers, 'sec-fetch-mode');
          assert_equals(headers['sec-fetch-mode'], 'same-origin');
        });
  }, 'module script');
  </script>
</html>
